requirements.txt
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
requests>=2.25.0
pyyaml>=6.0
pytest>=6.2.5
pytest-cov>=2.12.0
rich>=10.0.0
pytz>=2021.1
scikit-learn>=1.0.0
yfinance>=0.1.70
